requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
google-generativeai==0.3.2
chromadb==0.4.18
qdrant-client==1.11.1
sentence-transformers==2.2.2
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.0
python-dotenv==1.0.0
